# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API для Telegram Mini App по разделению счетов",
        "title": "BillSplitter API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:5001",
    "basePath": "/",
    "paths": {
...
        "/meets": {
            "get": {
                "security": [
                    {
                        "SessionAuth": []
                    }
                ],
                "description": "ID юзера берется из сессии",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "meet"
                ],
                "summary": "Получение персонального списка ивентов",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/internal_transport_http_meet.Event"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/billsplitter-monolith_internal_utils_http.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/billsplitter-monolith_internal_utils_http.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "SessionAuth": []
                    }
                ],
                "description": "Создает новый мит с участниками",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "meet"
                ],
                "summary": "Создает новый мит",
                "parameters": [
                    {
                        "description": "Данные мита",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_transport_http_meet.CreateEventRq"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/billsplitter-monolith_internal_utils_http.ResponseID"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/billsplitter-monolith_internal_utils_http.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/billsplitter-monolith_internal_utils_http.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/meets/{id}": {
            "get": {
                "security": [
                    {
                        "SessionAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "meet"
                ],
                "summary": "Получение инфы по ивенту",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID мита",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_transport_http_meet.Event"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/billsplitter-monolith_internal_utils_http.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/billsplitter-monolith_internal_utils_http.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/billsplitter-monolith_internal_utils_http.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "billsplitter-monolith_internal_utils_http.ErrorResponse": {
            "type": "object",
            "properties": {
                "errorMessage": {
                    "type": "string"
                }
            }
        },
        "billsplitter-monolith_internal_utils_http.ResponseID": {
            "type": "object",
            "properties": {
                "ID": {
                    "type": "integer"
                }
            }
        },
...
        "internal_transport_http_meet.CreateEventRq": {
            "type": "object",
            "properties": {
                "date": {
                    "type": "string"
                },
                "members": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "internal_transport_http_meet.Event": {
            "type": "object",
            "properties": {
                "ID": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "created_by_user_id": {
                    "type": "integer"
                },
                "deleted_at": {
                    "type": "string"
                },
                "event_date": {
                    "type": "string"
                },
                "members": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_transport_http_meet.Member"
                    }
                },
                "name": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "internal_transport_http_meet.Member": {
            "type": "object",
            "properties": {
                "ID": {
                    "type": "integer"
                },
                "user_id": {
                    "type": "integer"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "SessionAuth": {
            "description": "Сессионный токен. Значение из /auth/login/telegram передавайте через Authorize -\u003e X-Session-ID header.",
            "type": "apiKey",
            "name": "X-Session-ID",
            "in": "header"
        }
    }
}